requirements.txt
imgaug==0.4.0
kornia==0.6.12
matplotlib==3.6.2
numpy==1.23.4
opencv_python==4.6.0.66
Pillow==9.2.0
PyYAML==6.0
rawpy==0.18.1
scikit_image==0.19.3
scikit_learn==1.2.1
scipy==1.10.1
tifffile==2022.10.10
timm==0.3.2
tqdm==4.64.1
